requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.42.0
PyGithub==2.5.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
